+++ CHANGELOG.md
@@ -14,6 +14,7 @@ ENHANCEMENTS:
 FEATURES:
 * k8s: add `instance_template.name` attribute to `node group` resource and data source
 * k8s: add `instance_template.ipv4_dns_records`, `instance_template.ipv6_dns_records` attributes to `node group` resource and data source
+* k8s: add `instance_template.labels` attribute to `node group` resource and data source
 * greenplum: add fields `greenplum_config` and `pooler_config` in `resource_yandex_mdb_greenplum_cluster`
 * **New Resource:** `yandex_serverless_container_iam_binding`
 * **New Resource:** `yandex_ydb_database_iam_binding`